Update 2000-create-repo.md
Newly created git repositories have a 'master' branch, not a 'main' branch, yes? In my case, I had a 'master' branch. If that's not the convention, at least specify that some repo's will have a default 'master' branch. 

Or configure the CDK code to create a repo with a 'main' branch.